OSIS-4855 Move Technical exception from business exception
