+++ education_group/views/general_information/publish.py
@@ -6,8 +6,9 @@ from django.views.decorators.http import require_http_methods
 
 from base.views.common import display_error_messages, display_success_messages
 from education_group.ddd import command
-from education_group.ddd.domain.exception import PublishCommonAdmissionConditionException, \
-    PublishCommonPedagogyException
+from education_group.ddd.service.write.publish_common_pedagogy_service import PublishCommonPedagogyException
+from education_group.ddd.service.write.publish_common_admission_conditions_service import \
+    PublishCommonAdmissionConditionException
 from education_group.ddd.service.write import publish_common_pedagogy_service, \
     publish_common_admission_conditions_service
